tweaks to teleport gui, should fix some odd behavior on servers.
Also fixes an errored message in trainer edit gui.

+++ Pokecube Core/src/main/java/pokecube/core/network/PokecubePacketHandler.java
@@ -206,7 +206,7 @@ public class PokecubePacketHandler
                             }
                             else if (channel == TELEPORT)
                             {
-                                int index = buffer.readByte();
+                                int index = buffer.readInt();
                                 TeleportHandler.setTeleIndex(player.getCachedUniqueIdString(), index);
                             }
                         }
